Tilføjet/slettet og virker: - ApplicationMVC Package > PlayerController.java & PlayerService.java - Slettet PlayerConfig.java & PlayerRepository.java - RestAPIManager.java
