+++ src/main/java/com/gamersnation/gamersnationApplication/RestAPIManagers/RestAPIManagerRepository.java
@@ -1,16 +1,18 @@
-package com.gamersnation.gamersnationApplication.player;
+package com.gamersnation.gamersnationApplication.RestAPIManagers;
 
+import com.gamersnation.gamersnationApplication.player.Player;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
+
 import java.util.Optional;
 
-@RestController
+@Component
 @Repository
-public interface PlayerRepository extends JpaRepository<Player, String> {
+public interface RestAPIManagerRepository extends JpaRepository<Player, String> {
 
     Optional<Player> findPlayerByPuuid(String puuid);
 
@@ -18,6 +20,5 @@ public interface PlayerRepository extends JpaRepository<Player, String> {
 
     @Query(value = "SELECT name FROM Player")
     ArrayList<String> findALlPlayerNames();
-
 
 }
