--- src/main/java/com/gamersnation/gamersnationApplication/player/PlayerConfig.java
@@ -1,28 +1,0 @@
-package com.gamersnation.gamersnationApplication.player;
-
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-//Ændre navn til DBManeger
-@Configuration
-public class PlayerConfig {
-
-    @Bean
-    CommandLineRunner commandLineRunner(PlayerRepository repository) {
-        return args -> {
-            Player player = new Player(
-                    "Hamed",
-                    "Hamedxoxo",
-                    true,
-                    1,
-                    "Bronze",
-                    1,
-                    1,
-                    1,
-                    1
-            );
-            repository.save(player);
-        };
-    }
-}
