+++ src/main/resources/GUI.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<VBox xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gamersnation.gamersnationApplication.player.PlayerController">
+<VBox xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gamersnation.gamersnationApplication.ApplicationMVC.PlayerController">
     <GridPane>
         <Label styleClass="" fx:id="signUpLbl" text="Summoner Name" GridPane.columnIndex="1" GridPane.rowIndex="1"/>
         <TextField fx:id="summonerNameField" maxWidth="150" GridPane.columnIndex="1" GridPane.rowIndex="2"/>
